Create Run Salmon
The uniquely mapped reads from RNA-seq either fit in negative binomial or in the Gaussian/normal/Laplace-Gaussian distribution. They key question is how and where to “put” the multimapping sequencing reads. In the midst of piled up uniquely mapped reads, let’s imagine the coverage over the 3’UTR where the alignment appears gapped due unmapped multimappers. Since the multimappers are from unknown source and gapped alignments reflect the “incomplete data scenario”, we should not use any of the parametrical distribution models (i.e. known parameters most likely to give rise to the observed pattern e.g. maximum likelihood estimations) to fill the gaps. If we somehow find a way to “complete” the data by filling the gaps correctly by multimappers, the we can use the maximum likelihood methods to estimate the expression of a given transcript. The expectation maximization (EM) algorithm is a natural generalization of maximum likelihood estimation to the incomplete data case. The beauty of EM is that it can deal with the incomplete datasets showing the both binomial and Gaussian distributions. Thus, using the mixture models to identify the parameters that would be appropriate to assign the multimappers, then calculate the expected coverage for each data point (E), and lastly re-estimating the parameters (M) by plugging the significant multimappers over the individual transcripts should work fine.  Finally, the “E” and “M” steps are repeated until the convergence is achieved, meaning that we can then apply the Gaussian/binomial distribution to estimate the expression. In general words, The EM algorithm predicts the parameters (genomic locations in this case) that are most likely to be the source of multimappers, it then keeps plugging the multimappers to those locations until the overall coverage on each data point follows the distribution models that RNA-seq datasets generally follow.
Salmon run maps the sequencing reads based on the in-built EM algorithms