Update and rename Seurat V2 clustering (2016) to Seurat V2 clustering of preimplantation RNA-seq data(2016)
Codes for generating Figure 1, and Supplementary Figure 1 for selection arena  paper.
scRNA-seq data for petropolous et al., 2016, was downloaded in the read count expression matrices for E3-E7 from the ArrayExpress: E-MTAB-3929. It was then clustered using Seurat package
The above step was repeated for the Yan et al., 2013 data which was also publicly available. This was done independently, and also together with the merged analysis of 
#Petropolous et al., 2016